+++ src/app/profile/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 import { getProfile } from '@/app/profile/data/get-profile';
 import { Profile } from '@/app/profile/profile';
-import { verifySession } from '@/lib/permissions';
+import { verifySession } from '@/lib/dal';
 
 export const metadata: Metadata = {
     title: 'My Account',
