+++ src/app/user/create/page.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { CreateUser } from '@/app/user/create/create-user';
+import { verifySession } from '@/lib/dal';
 import { isUserAdministrator } from '@/lib/definitions';
-import { verifySession } from '@/lib/permissions';
 
 export const metadata: Metadata = {
     title: 'Create User',
